Add files via upload
We have one pipeline including some deep learning models. We are using a dataset of images that consists of clean and noisy images. For some use cases we want to pass only noisy images into the pipeline. So first we need to classify whether the image has a noise or not. so we are going to create one binary classification model using cnn.